Rename Panko::Impl → Panko::Engine; move lib/panko/impl → lib/panko/engine
The Impl namespace was a historical artifact from when this code was a
direct port of the C extension. Now that it is idiomatic Ruby, Engine
better reflects its role: the internal hot-path machinery that powers
serialization, distinct from the public Panko::* API surface.

- Rename lib/panko/impl/ → lib/panko/engine/ (git mv)
- Rename spec/unit/panko/impl/ → spec/unit/panko/engine/ (git mv)
- Replace all Panko::Impl → Panko::Engine references across lib/, spec/,
  benchmarks/, and panko_serializer.rb
- Update CLAUDE.md to reflect new paths and namespace
- Add .DS_Store to .gitignore
