+++ lib/panko/array_serializer.rb
@@ -45,7 +45,7 @@ Please pass valid each_serializer to ArraySerializer, for example:
     private
 
     def serialize_with_writer(subjects, writer)
-      srz = Panko::Impl::Serializer.new(@descriptor)
+      srz = Panko::Engine::Serializer.new(@descriptor)
       srz._serialize_many(subjects.to_a, writer)
       writer
     end
