+++ lib/panko/engine/attributes_writer/hash_writer.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-module Panko::Impl::AttributesWriter
-  class PlainWriter
+module Panko::Engine::AttributesWriter
+  class HashWriter
     def write_attributes(object, descriptor, writer)
       descriptor.attributes.each do |attr|
-        value = object.public_send(attr.name_sym)
+        value = object[attr.name]
         writer.push_value(value, attr.name_for_serialization)
       end
     end
