+++ lib/panko/engine/attributes_writer/plain_writer.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-module Panko::Impl::AttributesWriter
-  class HashWriter
+module Panko::Engine::AttributesWriter
+  class PlainWriter
     def write_attributes(object, descriptor, writer)
       descriptor.attributes.each do |attr|
-        value = object[attr.name]
+        value = object.public_send(attr.name_sym)
         writer.push_value(value, attr.name_for_serialization)
       end
     end
