+++ spec/unit/panko/serializer_spec.rb
@@ -181,9 +181,9 @@ describe Panko::Serializer do
         serializer = serializer_class.new
         mock_object = double("object")
 
-        # Stub Panko::Impl::Serializer to avoid dependencies on the object's attributes
+        # Stub Panko::Engine::Serializer to avoid dependencies on the object's attributes
         mock_impl_srz = double("srz", serialize_one: nil)
-        allow(Panko::Impl::Serializer).to receive(:new).and_return(mock_impl_srz)
+        allow(Panko::Engine::Serializer).to receive(:new).and_return(mock_impl_srz)
 
         # First call should work
         expect { serializer.serialize(mock_object) }.not_to raise_error
